+++ Source/Components/Graphics/Mesh.cpp
@@ -301,7 +301,7 @@ void Mesh::OnEditorInspect()
 				}
 				if (ImGui::BeginCombo(label.c_str(), ((texture) ? texture->GetPath().LocalPath.c_str() : "")))
 				{
-					for (int n = 0; n < Textures.size(); n++)
+					for (size_t n = 0; n < Textures.size(); n++)
 					{
 						if (ImGui::Selectable(Textures[n].LocalPath.c_str(), false))
 						{
